+++ RenderKit/openglatomshader.h
@@ -61,8 +61,8 @@ private:
   GLuint _program;
   std::vector<std::vector<std::shared_ptr<RKRenderStructure>>> _renderStructures;
 
-  std::vector<std::vector<int>> _numberOfIndices;
-  std::vector<std::vector<int>> _numberOfDrawnAtoms;
+  std::vector<std::vector<size_t>> _numberOfIndices;
+  std::vector<std::vector<size_t>> _numberOfDrawnAtoms;
 
   std::vector<std::vector<GLuint>> _vertexBuffer;
   std::vector<std::vector<GLuint>> _indexBuffer;
